Add tool for reviewing per-package changes
In order to review a release candidate PR, especially a large one, it is
best to cross-check the new changelog entries in the PR with the set of
changes that were made since the previous release. In the context of a
monorepo such as this one, we must consider packages when determing the
set of changes, as it's the packages we are publishing, not the entire
repo.

This is complicated by the fact that it is technically possible to issue
partial releases — that is, when the scope of a new release is decided,
it is possible to include only a subset of the total set of changed
packages. More concretely, if package A, B, and C are changed, but only
B is included in a release, then a future release will include A and/or
C. When _that_ release occurs, care must be taken to exclude the changes
in B which have already been released from the changelog.

In this kind of scenario, the best way to verify that the changelog is
being updated appropriately is to look at each package individually and
view the changes that have taken place since its previous release (not
the previous release of the whole monorepo).

This tool aids with such a comparison. Given the name of a workspace
such as `@metamask/address-book-controller`, it will compare the tag for
its latest version and HEAD and allow you to either view a set of
commits that have occurred in this timeframe along with their diffs, or
view the consolidated diff agnostic of commits. Note that `git` has been
customized to _not_ run its output through a pager, which means you'll
have to do this yourself:

    scripts/changes-since-last-release.ts --format per-commit @metamask/approval-controller | less -r
    scripts/changes-since-last-release.ts --format together @metamask/transaction-controller | less -r